+++ server/core_account.go
@@ -17,6 +17,10 @@ package server
 import (
 	"context"
 	"database/sql"
+	"encoding/json"
+	"github.com/heroiclabs/nakama/console"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"strconv"
 	"strings"
 	"time"
@@ -301,6 +305,104 @@ func UpdateAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, userID u
 	}
 
 	return nil
+}
+
+func ExportAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, userID uuid.UUID) (*console.AccountExport, error) {
+	// Core user account.
+	account, _, err := GetAccount(ctx, logger, db, nil, userID)
+	if err != nil {
+		if err == ErrAccountNotFound {
+			return nil, status.Error(codes.NotFound, "Account not found.")
+		}
+		logger.Error("Could not export account data", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+
+	// Friends.
+	friends, err := GetFriendIDs(ctx, logger, db, userID)
+	if err != nil {
+		logger.Error("Could not fetch friend IDs", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+
+	// Messages.
+	messages, err := GetChannelMessages(ctx, logger, db, userID)
+	if err != nil {
+		logger.Error("Could not fetch messages", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+
+	// Leaderboard records.
+	leaderboardRecords, err := LeaderboardRecordReadAll(ctx, logger, db, userID)
+	if err != nil {
+		logger.Error("Could not fetch leaderboard records", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+
+	groups := make([]*api.Group, 0)
+	groupUsers, err := ListUserGroups(ctx, logger, db, userID, 0, nil, "")
+	if err != nil {
+		logger.Error("Could not fetch groups that belong to the user", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+	for _, g := range groupUsers.UserGroups {
+		groups = append(groups, g.Group)
+	}
+
+	// Notifications.
+	notifications, err := NotificationList(ctx, logger, db, userID, 0, "", nil)
+	if err != nil {
+		logger.Error("Could not fetch notifications", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+
+	// Storage objects where user is the owner.
+	storageObjects, err := StorageReadAllUserObjects(ctx, logger, db, userID)
+	if err != nil {
+		logger.Error("Could not fetch notifications", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+
+	// History of user's wallet.
+	walletLedgers, err := ListWalletLedger(ctx, logger, db, userID)
+	if err != nil {
+		logger.Error("Could not fetch wallet ledger items", zap.Error(err), zap.String("user_id", userID.String()))
+		return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+	}
+	wl := make([]*console.WalletLedger, len(walletLedgers))
+	for i, w := range walletLedgers {
+		changeset, err := json.Marshal(w.Changeset)
+		if err != nil {
+			logger.Error("Could not fetch wallet ledger items, error encoding changeset", zap.Error(err), zap.String("user_id", userID.String()))
+			return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+		}
+		metadata, err := json.Marshal(w.Metadata)
+		if err != nil {
+			logger.Error("Could not fetch wallet ledger items, error encoding metadata", zap.Error(err), zap.String("user_id", userID.String()))
+			return nil, status.Error(codes.Internal, "An error occurred while trying to export user data.")
+		}
+		wl[i] = &console.WalletLedger{
+			Id:         w.ID,
+			UserId:     w.UserID,
+			Changeset:  string(changeset),
+			Metadata:   string(metadata),
+			CreateTime: &timestamp.Timestamp{Seconds: w.CreateTime},
+			UpdateTime: &timestamp.Timestamp{Seconds: w.UpdateTime},
+		}
+	}
+
+	export := &console.AccountExport{
+		Account:            account,
+		Objects:            storageObjects,
+		Friends:            friends.GetFriends(),
+		Messages:           messages,
+		Groups:             groups,
+		LeaderboardRecords: leaderboardRecords,
+		Notifications:      notifications.GetNotifications(),
+		WalletLedgers:      wl,
+	}
+
+	return export, nil
 }
 
 func DeleteAccount(ctx context.Context, logger *zap.Logger, db *sql.DB, userID uuid.UUID, recorded bool) error {